all instructions parameterized by u64

+++ jolt-core/src/jolt/instruction/slt.rs
@@ -9,8 +9,9 @@ use crate::{
   utils::instruction_utils::chunk_and_concatenate_operands,
 };
 
+// Values are stored as twos complement
 #[derive(Copy, Clone, Default, Debug)]
-pub struct SLTInstruction(pub i64, pub i64);
+pub struct SLTInstruction(pub u64, pub u64);
 
 impl JoltInstruction for SLTInstruction {
   fn combine_lookups<F: PrimeField>(&self, vals: &[F], C: usize, _: usize) -> F {
@@ -53,7 +54,7 @@ impl JoltInstruction for SLTInstruction {
   }
 
   fn to_indices(&self, C: usize, log_M: usize) -> Vec<usize> {
-    chunk_and_concatenate_operands(self.0 as u64, self.1 as u64, C, log_M)
+    chunk_and_concatenate_operands(self.0, self.1, C, log_M)
   }
 }
 
@@ -76,10 +77,10 @@ mod test {
     for _ in 0..256 {
       let x = rng.next_u64() as i64;
       let y = rng.next_u64() as i64;
-      jolt_instruction_test!(SLTInstruction(x, y), (x < y).into());
+      jolt_instruction_test!(SLTInstruction(x as u64, y as u64), (x < y).into());
     }
     for _ in 0..256 {
-      let x = rng.next_u64() as i64;
+      let x = rng.next_u64();
       jolt_instruction_test!(SLTInstruction(x, x), Fr::zero());
     }
   }
